Switch CallStackTest and ValuesTest to use MatchMembers.
Fix matching of null members in MatchMember.

Refactor classes in the DebugEvaluationResult hierarchy into their own separate files.
