+++ src/Debugger/Impl/DebugActiveBindingEvaluationResult.cs
@@ -1,0 +1,22 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Common.Core;
+using Microsoft.R.Host.Client;
+using Newtonsoft.Json.Linq;
+using static System.FormattableString;
+
+namespace Microsoft.R.Debugger {
+    public class DebugActiveBindingEvaluationResult : DebugEvaluationResult {
+        public DebugActiveBindingEvaluationResult(DebugStackFrame stackFrame, string expression, string name)
+            : base(stackFrame, expression, name) {
+        }
+    }
+}
