+++ src/Debugger/Impl/DebugStackFrame.cs
@@ -93,9 +93,8 @@ namespace Microsoft.R.Debugger {
             return EvaluateAsync("base::environment()", fields: fields, cancellationToken: cancellationToken);
         }
 
-        public override string ToString() {
-            return Invariant($"{Call ?? "<null>"} at {FileName ?? "<null>"}:{(LineNumber != null ? LineNumber.ToString() : "<null>")}");
-        }
+        public override string ToString() =>
+            Invariant($"{EnvironmentName ?? Call ?? "<null>"} at {FileName ?? "<null>"}:{(LineNumber?.ToString() ?? "<null>")}");
 
         public override bool Equals(object obj) =>
             base.Equals(obj) || (obj as IEquatable<DebugStackFrame>)?.Equals(this) == true;
