+++ src/Debugger/Test/Match/MatchDebugStackFrames.cs
@@ -4,12 +4,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq.Expressions;
-using System.Reflection;
-using System.Text;
-using FluentAssertions.Formatting;
 using Microsoft.Common.Core.Test.Match;
-using static System.FormattableString;
 
 namespace Microsoft.R.Debugger.Test.Match {
     internal class MatchDebugStackFrames : IEnumerable<IEquatable<DebugStackFrame>> {
@@ -23,15 +18,28 @@ namespace Microsoft.R.Debugger.Test.Match {
             return GetEnumerator();
         }
 
-        public void Add(IEquatable<string> fileName, IEquatable<int> lineNumber, IEquatable<string> call) {
-            _frames.Add(new MatchMembers<DebugStackFrame>()
+        public void Add(IEquatable<DebugStackFrame> match) {
+            _frames.Add(match);
+        }
+
+        public void Add(IEquatable<string> fileName, IEquatable<int> lineNumber, IEquatable<string> call, IEquatable<string> environmentName) {
+            Add(new MatchMembers<DebugStackFrame>()
                 .Matching(x => x.FileName, fileName)
                 .Matching(x => x.LineNumber, lineNumber)
-                .Matching(x => x.Call, call));
+                .Matching(x => x.Call, call)
+                .Matching(x => x.EnvironmentName, environmentName));
+        }
+
+        public void Add(IEquatable<string> fileName, IEquatable<int> lineNumber, IEquatable<string> call) {
+            Add(fileName, lineNumber, call, MatchAny<string>.Instance);
         }
 
         public void Add(IEquatable<string> fileName, IEquatable<int> lineNumber) {
             Add(fileName, lineNumber, MatchAny<string>.Instance);
+        }
+
+        public void Add(SourceFile sourceFile, IEquatable<int> lineNumber, IEquatable<string> call, IEquatable<string> environmentName) {
+            Add(sourceFile.FilePath, lineNumber, call, environmentName);
         }
 
         public void Add(SourceFile sourceFile, IEquatable<int> lineNumber, IEquatable<string> call) {
